+++ tests/unit/rules/prefer-alphabetical-bundledDependencies.js
@@ -22,7 +22,7 @@ describe('prefer-alphabetical-bundledDependencies Unit Tests', function() {
       response.lintId.should.equal('prefer-alphabetical-bundledDependencies');
       response.lintType.should.equal('error');
       response.node.should.equal('bundledDependencies');
-      response.lintMessage.should.equal('Your bundledDependencies are not in alphabetical order. Please update the order.');
+      response.lintMessage.should.equal('Your bundledDependencies are not in alphabetical order. Please move semver after chalk.');
     });
   });