+++ tests/unit/rules/prefer-alphabetical-dependencies.js
@@ -22,7 +22,7 @@ describe('prefer-alphabetical-dependencies Unit Tests', function() {
       response.lintId.should.equal('prefer-alphabetical-dependencies');
       response.lintType.should.equal('error');
       response.node.should.equal('dependencies');
-      response.lintMessage.should.equal('Your dependencies are not in alphabetical order. Please update the order.');
+      response.lintMessage.should.equal('Your dependencies are not in alphabetical order. Please move semver after chalk.');
     });
   });